+++ app/routes.py
@@ -4,9 +4,9 @@ from sdx_gcp.errors import DataError
 
 from app.definitions.spec import Template
 from app.definitions.data import Data, SurveyMetadata, Identifier, PrepopData, ListCollector, PCK
-from app.pck_managers.looped import get_looping
-from app.pck_managers.flat import get_pck
-from app.pck_managers.prepop import get_prepop
+from app.controllers.looped import get_looping
+from app.controllers.flat import get_pck
+from app.controllers.prepop import get_prepop
 
 
 logger = get_logger()
