+++ app/transformers/looped.py
@@ -3,7 +3,6 @@ from typing import Literal
 from app.definitions.data import SurveyMetadata
 from app.definitions.spec import BuildSpec, ParseTree
 from app.formatters.looping_formatter import LoopingFormatter
-from app.transform.interpolate import interpolate
 from app.transform.populate import resolve_value_fields
 from app.transformers.standard import SpecTransformer
 
@@ -24,7 +23,7 @@ class LoopedPckSpecTransformer(SpecTransformer[SurveyMetadata, LoopingFormatter]
             template = "template"
 
         if 'transforms' in build_spec:
-            parse_tree: ParseTree = interpolate(build_spec[template], build_spec["transforms"])
+            parse_tree: ParseTree = self._executor.interpolate(build_spec[template], build_spec["transforms"])
         else:
             parse_tree: ParseTree = build_spec[template]
         return resolve_value_fields(parse_tree)
