+++ tests/integration/berd/test_berd_long.py
@@ -2,7 +2,7 @@ import json
 import unittest
 
 from app.definitions.data import SurveyMetadata, SPP, ImageResponse
-from app.pck_managers.looped import get_looping
+from app.controllers.looped import get_looping
 from tests.integration.looped import read_submission_data
