+++ tests/integration/mapped/test_qcas.py
@@ -2,7 +2,7 @@ import os
 import unittest
 
 from app.definitions.data import Data, SurveyMetadata, PCK
-from app.pck_managers.flat import get_pck
+from app.controllers.flat import get_pck
 from tests.integration.mapped import read_submission_data, are_equal
