+++ tests/unit/berd/test_berd_transformer.py
@@ -5,7 +5,7 @@ from app.berd.collect_items import collect_list_items
 from app.berd.convert_data import extract_answers, convert_to_spp
 from app.berd.definitions import Answer, SPP
 from app.definitions.data import SurveyMetadata
-from app.pck_managers.looped import get_looping
+from app.controllers.looped import get_looping
 
 
 class BERDTransformerTests(unittest.TestCase):
